backend/requirements.txt
jaseci
python-dotenv
requests
streamlit
gitpython
tree_sitter
flask
